logstash-initContainer/requirements.txt
Cerberus==1.3.5
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.1
